Multiple source directory support for sbt-{less,js,mustache}.
You can not specify multiple source directories in Java CLASSPATH style:

    (sourceDirectories in (Compile, FooKeys.foo)) <<= (sourceDirectory in Compile) {
      srcDir =>
        Seq(
          srcDir / "first" / "path",
          srcDir / "second" / "path",
          // etc ...
        )
    }

Files in earlier source directories can "shadow" files in later directories if they have the same relative path. For example, in the example above the file:

    .../.../.../first/path/foo.js

would shadow:

    .../.../.../second/path/foo.js

When a file is added to the source graph, the plugin checks to see if it is shadowed by another file. If it is, the shadowing file is added to the graph instead. This allows developers to override individual files in a library without copying and editing the contents of files that depend on them.