Make lessc command OS-independent
Previously, the lessc command issued to a subprocess when useCommandLine = true
omitted a file extention, implying that it was expected to be a reference to an
executeable file in a Unix-like environment.  The windows version of lessc
includes the file extension .cmd, so the command execution omitting it would
fail with an IOException.